requirements.txt
# SCLU Trading System - Production Dependencies

# Core trading and backtesting
backtrader>=1.9.76.123
kiteconnect>=4.1.0

# Data analysis and processing
pandas>=1.3.0
numpy>=1.21.0

# Configuration and utilities
PyYAML>=6.0
python-dotenv>=0.19.0

# Logging and monitoring
colorlog>=6.6.0

# Date and time handling
python-dateutil>=2.8.2

# HTTP requests and API communication
requests>=2.25.1
urllib3>=1.26.0

# Optional: Enhanced data visualization
matplotlib>=3.5.0
plotly>=5.0.0

# Optional: Performance optimization
numba>=0.56.0

# Optional: Database support
# SQLAlchemy>=1.4.0
# psycopg2-binary>=2.9.0

# Optional: Advanced analytics
# scipy>=1.7.0
# scikit-learn>=1.0.0
# ta-lib>=0.4.24
